requirements.txt
scapy==2.4.3
urllib3==2.0.7
netaddr==0.7.19
netifaces==0.10.9
miniupnpc==2.0.2
dnspython==1.16.0
requests==2.31.0
xeger==0.3.5
waitress==1.4.3
Flask==1.1.1
twisted==22.8.0
asyncssh==2.1.0
service-identity==21.1.0
pycryptodomex
netfilterqueue